+++ R/check-cran.r
@@ -132,13 +132,6 @@ check_cran <- function(pkgs, libpath = file.path(tempdir(), "R-lib"),
     results
   }
 
-  if (getRversion() <= '2.15.2' && threads >= length(pkgs)) {
-    threads <- max(length(pkgs) - 1L, 1L)
-    message("Reducing number of threads to ", threads,
-      " (number of packages to check minus one) due to a bug in mclapply in",
-      " R <= 2.15.2")
-  }
-
   results <- mclapply(seq_along(pkgs), check_pkg, mc.preschedule = FALSE,
     mc.cores = threads)