+++ R/zzz.r
@@ -46,8 +46,7 @@ NULL
     devtools.desc.author = '"First Last <first.last@example.com> [aut, cre]"',
     devtools.desc.license = "What license is it under?",
     devtools.desc.suggests = NULL,
-    devtools.desc = list(),
-    github.user = "hadley"
+    devtools.desc = list()
   )
   toset <- !(names(op.devtools) %in% names(op))
   if(any(toset)) options(op.devtools[toset])